move files to match class names
