+++ lib/timing_attack.rb
@@ -1,10 +1,10 @@
 require 'typhoeus'
 require 'ruby-progressbar'
 require "timing_attack/version"
-require "timing_attack/discoverer"
+require "timing_attack/brute_forcer"
 require "timing_attack/grouper"
 require "timing_attack/test_case"
-require "timing_attack/cli_attacker"
+require "timing_attack/enumerator"
 
 module TimingAttack
 end
