+++ lib/timing_attack/brute_forcer.rb
