+++ lib/timing_attack/enumerator.rb
